requirements.txt
streamlit>=1.29.0
groq>=0.4.0
gTTS>=2.3.0
moviepy>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
